Project 8: add support for functions
Adds the functionality to translate function initialisations, calling and returns.

Changes binary expressions to store temp variables in the static variable space in RAM in place of the temp memory segment of a namespace. This separates out the memory used by the functions themselves and temp allocation of memory required for computation, reducing the case of accidental overriding of memory.